==> scripts/03_meninges_R_analysis/02_variance.Rmd <==
---
title: "Meningitis: Meninges Tissue"
subtitle: "Variance"
author: "Kennedi Todd"
date: "'r Sys.Date()'"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup
## Working directory
```{r working_directory}
knitr::opts_knit$set(root.dir = ".")
```

## Libraries
```{r libraries}
library(corrplot)          # corrplot()
library(dplyr)             # %>%
library(edgeR)             # load DGEList object
library(ggcorrplot)        # ggcorrplot()
library(gplots)            # heatmap.2()
library(limma)             # voomWithQualityWeights()
library(mvIC)              # mvForwardStepwise()
library(variancePartition) # fitExtractVarPartModel()
```

## User defined variables
```{r set_variables}
tissue <- "meninges"
folder <- paste0(tissue, "_no_VaD")
groups <- c("control","meningitis")
group_colors <- c("orange","cornflowerblue")
sex_colors <- c("green","purple")
```

## Save functions
These functions with help simultaneously save plots as a png, pdf, and tiff 
file.
```{r save_functions}
saveToPDF <- function(...) {
    d = dev.copy(pdf,...)
    dev.off(d)
}
```

## Load data
```{r load_data}
dge.filtered.norm <- readRDS(paste0("../../rObjects/",
                                    folder,
                                    "_DGEList_filtered_normalized.rds"))
```

# Cell Markers
## Extract metadata
```{r}
# get meta
meta <- as.data.frame(dge.filtered.norm$samples)

# keep variables with all fields known
variables <- c("group",
               "sex",
               "lib.size",
               "braak",
               "thal",
               "age_at_death",
               "brain_wt",
               "RIN",
               "star_unique_mapped")
meta <- meta %>% select(all_of(variables))

# log2 transform lib.size
meta$log2.lib <- log2(meta$lib.size)
meta$lib.size <- NULL

# check and change classes
meta$braak <- factor(meta$braak)
meta$thal <- factor(meta$thal)
meta$RIN <- as.numeric(meta$RIN)
meta$star_unique_mapped <- as.numeric(gsub("%","",meta$star_unique_mapped))
sapply(meta, class)
```

## Markers
- Astrocytes: GFAP, GJA1, CLU, AQP4
- B cells & Plasma Cells: CD19, MS4A1, CD79A, IGHA, SDC1
- Choroid: TTR
- Endothelial: FLT1, PECAM1, VWF \
- Fibroblasts: COL1A1, COL1A2, LUM, DCN \
- Macrophages/Microglia: MRC1, MKI67, C1QB \
- Mast cells: KIT, FCER1A \
- Monocytes/DCs/Neutrophils: ITGAX, CD14, CD209A, S100A9 \
- Neurons: GAD1, GAD2, SNAP25, SYT1, MEG3
- NPCs: SOX2, NES
- NK cells: NKG7
- Oligodendrocytes: MBP
- OPCs: OLIG1, TNR
- Pericytes & SMCs: MYH11, ACTA2, VTN \
- RBCs - HBB, HBA1, HBA2 \
- Schwann cells: MPZ \
- T cells: CD3E, CD3D, CD8A \
```{r}
# choose cell type markers
markers <- c("GFAP", "CD3E", "TTR", "VWF", "COL1A2", "MRC1", "KIT", "ITGAX", 
             "SNAP25", "NES", "NKG7", "MBP", "TNR", "MYH11", "HBB", "MPZ")
length(markers)

# print markers not in annotation (could have alternative name)
markers[!markers %in% rownames(dge.filtered.norm)]

# get counts
cpm <- round(as.data.frame(cpm(dge.filtered.norm$counts, log = TRUE)), 
             digits = 2)

# filter counts by marker
cpm <- cpm[rownames(cpm) %in% markers,]

# switch cols and rows
cpm <- t(cpm)

# add to meta
all.equal(rownames(meta), rownames(cpm))
meta <- cbind(meta, cpm)
```

## Significance
```{r stats}
# initialize data frame
variable_numbers <- c(5:24)
variable_names <- colnames(meta)[variable_numbers]
tests <- c("control_mean","meningitis_mean","control_stdv","meningitis_stdv",
           "control_shapiro_pval","meningitis_shapiro_pval","levene_na_omit",
           "t_test_t","t_test_pval","wilcoxon_pval")
stats <- matrix(0,
                nrow = length(variable_numbers),
                ncol = length(tests),
                dimnames = list(variable_names, tests))
stats <- as.data.frame(stats)

# Loop through variables
for (i in 1:length(variable_names)) {
  
  # Subset values by group
  values <- as.numeric(meta[, variable_names[i]])
  group <- factor(meta$group)
  control_values <- meta[meta$group == "control", variable_names[i]]
  meningitis_values <- meta[meta$group == "meningitis", variable_names[i]]
  
  # Welch two sample t-test
  t_test <- t.test(control_values, meningitis_values)
  t_test_t <- t_test$statistic
  t_test_pval <- t_test$p.value
  
  # Add to table
  stats[i, "control_mean"] <- round(as.numeric(t_test$estimate[1]),4)
  stats[i, "meningitis_mean"] <- round(as.numeric(t_test$estimate[2]),4)
  stats[i, "t_test_t"] <- round(t_test_t,4)
  stats[i, "t_test_pval"] <- round(t_test_pval,4)
  
  # Shapiro-Wilk test, normality test
  control_shapiro <- shapiro.test(control_values)
  meningitis_shapiro <- shapiro.test(meningitis_values)
  
  # Add to table
  stats[i, "control_shapiro_pval"] <- round(control_shapiro$p.value,4)
  stats[i, "meningitis_shapiro_pval"] <- round(meningitis_shapiro$p.value,4)
  
  # Variance and standard deviation
  control_var <- var(control_values)
  meningitis_var <- var(meningitis_values)
  
  # Add to table 
  stats[i, "control_stdv"] <- round(sqrt(control_var),4)
  stats[i, "meningitis_stdv"] <- round(sqrt(meningitis_var),4)
  
  # Two-sample Wilcoxon / Mann-Whitney
  wil <- wilcox.test(control_values, meningitis_values)
  
  # Add to table
  stats[i, "wilcoxon_pval"] <- round(wil$p.value, 4)
  
  # Levene's test
  levTest <- car::leveneTest(values ~ group)
  lev <- levTest$`Pr(>F)`
  lev_naomit <- as.numeric(na.omit(lev))
  
  # Add to table
  stats[i, "levene_na_omit"] <- round(lev_naomit, 4)
}

remove(control_shapiro,meningitis_shapiro,
       t_test, wil, levTest,cpm)
```

# Correlation
## Spearman correlation
- Spearman's correlation can be used to analyzed categorical data, as well as continuous data, because it uses ranks instead of assumptions of normality ().
plotCorrMatrix(cor(info))
MRC1, MKI67, C1QB
```{r CCA_heatmap}
# formula
form <- ~0 + group + sex + log2.lib + braak + thal + age_at_death + brain_wt +
  RIN + star_unique_mapped

# correlation plot
model.matrix(form, data = meta) %>% 
  cor(method = "spearman") %>% 
  ggcorrplot(show.diag = FALSE, type = "lower", lab = TRUE, lab_size = 2)

# save
path <- paste0("../../results/", folder, "/variance/spearman_correlation_heatmap")
saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
```

## CCA heatmap
```{r}
c <- canCorPairs(form, meta)

corrplot(c,
         type = "lower",
         is.corr = FALSE,
         order = "hclust",
         col = colorRampPalette(c("blue4", "deepskyblue", "white"))(100),
         addCoef.col = TRUE,
         method = "shade",
         tl.col = "black",
         tl.cex = 0.8,
         number.cex = 0.7,
         diag = TRUE)

# save
path <- paste0("../../results/", folder, "/variance/canonical_correlation_analysis_heatmap")
saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
```

# Variance partition
## Variance model
```{r variance_partition, message=FALSE, eval=FALSE}
# variables to choose from
geneExpr <- as.matrix(dge.filtered.norm$counts)
colnames(meta)

# age is usually a continuous so model it as a fixed effect "age"
# group is categorical, so model them as random effects "(1|group)"
form <- ~ (1|group) + (1|sex) + (1|braak) + (1|thal) + age_at_death + brain_wt + RIN

varPart <- fitExtractVarPartModel(geneExpr, form, meta)
vp <- sortCols(varPart)
saveRDS(vp, paste0("../../rObjects/", folder, "_var_par.rds"))
```

## Violins
```{r variance_violins, eval=FALSE}
# plot
plotVarPart(vp)

# save
path <- paste0("../../results/", folder, "/variance/variance_violins")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 6)
```

## Percent bars
```{r variance_percent_bars}
# plot
plotPercentBars(vp[1:10,])

# save
path <- paste0("../../results/", folder, "/variance/variance_gene_percent_bars")
saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
```

## Group var
```{r variance_group}
# sort genes based on variance explained by group
varPart.df <- as.data.frame(vp)
order.varPart.df <- varPart.df[order(varPart.df$group, decreasing = TRUE),]
head(order.varPart.df["group"], 10)
```

## Sex var
```{r variance_group}
# sort genes based on variance explained by group
order.varPart.df <- varPart.df[order(varPart.df$sex, decreasing = TRUE),]
head(order.varPart.df["sex"], 10)
```

## Lib size var
```{r variance_group}
# sort genes based on variance explained by group
order.varPart.df <- varPart.df[order(varPart.df$log2.lib, decreasing = TRUE),]
head(order.varPart.df["log2.lib"], 10)
```

## Thal var
```{r variance_group}
# sort genes based on variance explained by group
order.varPart.df <- varPart.df[order(varPart.df$thal, decreasing = TRUE),]
head(order.varPart.df["thal"], 10)
```

# Finding best model
variancePartition::voomWithDreamWeights() is the same as limma::voom(), except that it allows random effects in the formula. limma::voomWithQualityWeights() includes observational and sample specific quality weights whereas voomWithDreamWeights only includes observational quality weights.

## Forward stepwise regression
```{r}
# get matrix of expression data
v <- voomWithDreamWeights(counts = dge.filtered.norm$counts,
                          formula = (~ 0 + group + sex),
                          data = dge.filtered.norm$samples,
                          BPPARAM = BiocParallel::SnowParam(cores),
                          plot = FALSE)
counts <- v$E

# set base formula
base <- ~ (1 | group) + (1 | sex)

# variable we wish to test
variables <- c("(1|braak)",
               "(1|thal)",
               "age_at_death",
               "brain_wt",
               "RIN")

# multivariate forward stepwise regression evaluated by multivariate BIC
bestModel <- mvForwardStepwise(exprObj = counts,
                               baseFormula = base,
                               data = meta,
                               variables = variables)
```

# Save
```{r}
all.equal(rownames(meta), rownames(dge.filtered.norm$samples))
dge.filtered.norm$samples <- cbind(dge.filtered.norm$samples, meta[,9:24])
saveRDS(dge.filtered.norm, paste0("../../rObjects/",folder, "_DGEList_cell_markers.rds"))
```

```{r}
sessionInfo()
```
